[reductors] rename lti to h2, move tf into interpolation and h2, logging improvements
